Implement rabbitmq test container
