+++ server/src/main/java/com/szampchat/server/event/rabbitmq/RabbitMqConnectionProperties.java
@@ -1,12 +1,14 @@
 package com.szampchat.server.event.rabbitmq;
 
 import lombok.Getter;
+import lombok.Setter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
 @Component
 @ConfigurationProperties("spring.rabbitmq")
 @Getter
+@Setter
 public class RabbitMqConnectionProperties {
     private String host = "localhost";
     private int port = 5672;
