+++ server/src/test/java/com/szampchat/server/PostgresTestContainer.java
@@ -1,0 +1,26 @@
+package com.szampchat.server;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+
+public interface PostgresTestContainer {
+    @Container
+    PostgreSQLContainer<?> postgres = new PostgreSQLContainer<>("postgres:16-alpine");
+
+    @DynamicPropertySource
+    static void configureProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.r2dbc.url", RdbcUrlUtil.getR2dbcUrl(postgres));
+        registry.add("spring.r2dbc.username", postgres::getUsername);
+        registry.add("spring.r2dbc.password", postgres::getPassword);
+    }
+
+    @Test
+    default void connectionEstablished() {
+        Assertions.assertTrue(postgres.isCreated());
+        Assertions.assertTrue(postgres.isRunning());
+    }
+}
