+++ server/src/test/java/com/szampchat/server/RabbitMQTestContainer.java
@@ -1,0 +1,28 @@
+package com.szampchat.server;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.containers.RabbitMQContainer;
+import org.testcontainers.junit.jupiter.Container;
+
+public interface RabbitMQTestContainer {
+    @Container
+    RabbitMQContainer rabbitmq = new RabbitMQContainer("rabbitmq:3.13-alpine");
+
+    @DynamicPropertySource
+    static void configureProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.rabbitmq.host", rabbitmq::getHost);
+        registry.add("spring.rabbitmq.port", rabbitmq::getAmqpPort);
+        registry.add("spring.rabbitmq.username", rabbitmq::getAdminUsername);
+        registry.add("spring.rabbitmq.password", rabbitmq::getAdminPassword);
+    }
+
+    @Test
+    default void connectionEstablished() {
+        Assertions.assertTrue(rabbitmq.isCreated());
+        Assertions.assertTrue(rabbitmq.isRunning());
+    }
+}
